First and foremost. Added jquery and bootstrap.js since SOMEONE forgot to commit those... UUUGHHGGHGH
Added a nifty results() function to polls that returns a list of tuples with (choicename, votes). Handy for templates
modified base.html to let add some more scripting at the end of the page
modified results.html to show a nifty animated barchart!!
Added a library that totally does the whole thing for me!
    * I'll look into raphael.js if we need more advanced stuff that what is already there.

Ok Perro, MAKE IT PRETTY