LESS is working.
Ok - so bootstrap assumes that you already know what you're doing with LESS and serves as just documentation NOT A
TUTORIAL.

Basically bootstrap.css (as we know it) is created by compiling bootstrap.less. bootstrap.less is made up of tons of
*.less files that together define all of the properties known in bootstrap.less.

Since we're using javascript less, our browser does the converstion from all of these less files into what's basically a
single css file.
Ideally django should be compiling our less files so that not every user has to download/compile on EVERY DAMNED
PAGELOAD. But for development this is fine since it only takes in chrome ~300ms on my machine.

We should probably have a *.less file outside of the boostrap directory that we modify and include so that our changes
don't get mixed up with bootstraps since we'll definitaly be making changes to the bootstrap.less. (The git version is
EXTREMELY bleeding edge, commented out progress-bars to get the stupid thing to compile)